Fix incorrect "field not initialized" error with try/else
Previously, due to a bug in the compiler's while/else handling code, the
following code would be rejected as not initializing all object fields:

```pony
actor Main
  var _s: (String | None)

  new create(env: Env) =>
    try
      _s = "".usize()?.string()
    else
      _s = None
    end
```

This commit fixes try/else handling in `refer_try` so that it matches the
general shape of `refer_if`, `refer_while`, and `refer_repeat`. It differs from
them because `try` isn't a scope and we want to push everything into the `then`
clause as it can fulfill initialization all on its own or if it doesn't handle
initialization, we can still get it from the try body and the else clause.

Closes #3283